+++ tests/integration/DisputeSearchTest.php
@@ -86,6 +86,36 @@ class DisputeSearchTest extends Setup
         $this->assertEquals($disputes[0]->protectionLevel, Braintree\Dispute::EFFORTLESS_CBP);
     }
 
+    public function testAdvancedSearch_byPreDisputeProgram_returnsDispute()
+    {
+        $collection = Braintree\Dispute::search([
+            Braintree\DisputeSearch::preDisputeProgram()->in([
+                Braintree\Dispute::VISA_RDR
+            ])
+        ]);
+
+        $disputes = $this->collectionToArray($collection);
+
+        $this->assertEquals(1, count($disputes));
+        $this->assertEquals($disputes[0]->preDisputeProgram, Braintree\Dispute::VISA_RDR);
+    }
+
+    public function testAdvancedSearch_forNonPreDisputes_returnsDisputes()
+    {
+        $collection = Braintree\Dispute::search([
+            Braintree\DisputeSearch::preDisputeProgram()->is(Braintree\Dispute::NONE)
+        ]);
+
+        $disputes = $this->collectionToArray($collection);
+        $preDisputePrograms = array_unique(array_map(function ($d) {
+            return $d->preDisputeProgram;
+        }, $disputes));
+
+        $this->assertGreaterThan(1, count($disputes));
+        $this->assertEquals(1, count($preDisputePrograms));
+        $this->assertTrue(in_array(Braintree\Dispute::NONE, $preDisputePrograms));
+    }
+
     public function testAdvancedSearch_byReceivedDateRange_returnsDispute()
     {
         $collection = Braintree\Dispute::search([